requirements.txt
alpaca-py>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
streamlit>=1.30.0
joblib>=1.3.0
ta>=0.10.0
matplotlib>=3.7.0
plotly>=5.15.0
azure-storage-blob>=12.0.0
azure-data-tables>=12.4.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
pyyaml>=6.0
